get_envirem_elev: Get elevation data from envirem

Get elevation layers from ENVIREM for present and past scenarios

get_envirem_elev(period, region, resolution, fmt, sv_dir = getwd(), load2r)

Arguments

period
Period. Possible values are: "current" ( 1960 - 1990),"holo" (Mid-Holocene ~ 6000 years ago),"lgm" (Last Glacial Maximum ~ 22000 years ago).

region
Geographic region. Posibles values are: "Africa", "Australia", "Eurasia", "Europe", "global", "NAmerica", "NewWorld", "Pacific", and, "SAmerica".

resolution
Resolution of layers. Posible values are "10arcmin","5arcmin","2.5arcmin", and, "30arcsec".

fmt
File format. Possible values are "bil" and "geotiff".

sv_dir
Path to the directory where the layers will be saved. Default is the working directory of the R session.

load2r
Logical. Load layers into R?

Details

For more details visit https://envirem.github.io/

References

Title P.O., Bemmels J.B. 2018. ENVIREM: an expanded set of bioclimatic and topographic variables increases flexibility and improves the performance of ecological niche modeling. Ecography. 41:291-307.

See also

Examples

if (FALSE) {
# For a compleate list of posible values of the function
# elev_data <- base::readRDS(file.path(system.file("extdata/rds",
#                                                   package = "ntbbox"),
#                                                  "envirem_elevation.rds"))

elev_eu_lgm_10min <- get_envirem_elev(period="lgm",
                                      region="Europe",
                                      resolution= "10arcmin",
                                      fmt ="geotiff",
                                      sv_dir = "~/Desktop",
                                      load2r=TRUE)
}